Use a map for conversions
Even though this results in slightly worse code for the TUI (needing to
convert the map to a list first) and an xToString function instead of a
proper Show instance, it still seems conceptually nicer to use a map
here.  We are obviously disambiguating different conversions by their
keybindings, so we might as well make this explicit in the code.